SubScreen System Created
The Screen Classes had a lot of work loaded into them. Now SubScreens are used to divy up parts of the screen. Screens are now basically controllers, and SubScreen views (MVC pattern)

SubScreens do not inherit/ implement anything common. Not enough is written for it be clear what the best approach is.