Auto kiln & auto forge now implemented
The copper workbench is the automated version of
the workbench. That factory class (AutoCrafter) now
supports Kiln & Forge. In the actual game, you
craft these seperately, meaning that there is now an
Iron Kiln & Steel Forge.

Alot of files were touched though because these things have
4 representations. 1 in Tile, 1 in Glyph, 1 in ItemIndex, and 1
in FactoryAutoCrafter which all had to be adjusted to work
with each other.